+++ webviewsrc/focustree.ts
@@ -4,7 +4,7 @@ import { difference, minBy } from "lodash";
 import { renderGridBoxCommon, GridBoxItem, GridBoxConnection } from "../src/util/hoi4gui/gridboxcommon";
 import { StyleTable, normalizeForStyle } from "../src/util/styletable";
 import { FocusTree, Focus, UpdateFocusPositionsMessage } from "../src/previewdef/focustree/schema";
-import { applyCondition, ConditionItem } from "../src/hoiformat/condition";
+import { applyCondition, ConditionItem, conditionItemToStringValue, conditionToString, stringValueToConditionItem } from "../src/hoiformat/condition";
 import { NumberPosition } from "../src/util/common";
 import { GridBoxType } from "../src/hoiformat/gui";
 import { toNumberLike } from "../src/hoiformat/schema";
@@ -177,9 +177,9 @@ function updateSelectedFocusTree(clearCondition: boolean) {
         if (conditions) {
             conditions.select.innerHTML = `<span class="value"></span>
                 ${conditionExprs.map(option =>
-                    `<div class="option" value='${option.scopeName}!|${option.nodeContent}'>${option.scopeName ? `[${option.scopeName}] ` : ''}${option.nodeContent}</div>`
+                    `<div class="option" value='${conditionItemToStringValue(option)}'>${conditionToString(option)}</div>`
                 ).join('')}`;
-            conditions.selectedValues$.next(clearCondition ? [] : selectedExprs.map(e => `${e.scopeName}!|${e.nodeContent}`));
+            conditions.selectedValues$.next(clearCondition ? [] : selectedExprs.map(conditionItemToStringValue));
         }
 
     } else {
@@ -672,25 +672,11 @@ window.addEventListener('load', tryRun(async function() {
         if (conditionsElement) {
             conditions = new DivDropdown(conditionsElement, true);
             
-            conditions.selectedValues$.next(selectedExprs.map(e => `${e.scopeName}!|${e.nodeContent}`));
+            conditions.selectedValues$.next(selectedExprs.map(conditionItemToStringValue));
             conditions.selectedValues$.subscribe(async (selection) => {
-                selectedExprs = selection.map<ConditionItem>(selection => {
-                    const index = selection.indexOf('!|');
-                    if (index === -1) {
-                        return {
-                            scopeName: '',
-                            nodeContent: selection,
-                        };
-                    } else {
-                        return {
-                            scopeName: selection.substring(0, index),
-                            nodeContent: selection.substring(index + 2),
-                        };
-                    }
-                });
-
+                selectedExprs = selection.map<ConditionItem>(stringValueToConditionItem);
                 setState({ selectedExprs });
-                
+
                 await buildContent();
                 retriggerSearch();
             });